sql: hook up new schema change to connExecutor
This commit adds support for execution for the statement and
pre-commit phases of the schema changer to the `connExecutor`.

Also adds a session/cluster setting
`sql.defaults.experimental_new_schema_changer.mode`, which is `'off'`
default. The setting takes on the following three values:

 * `'off'` - never use the new schema changer
 * `'on'` - use the new schema changer in supported, implicit transaction
 * `'unsafe_always'` - always try to use the new schema changer, return errors
   where not supported

This code also hooks up the planning logic for the above settings. As of this
commit, only `ALTER TABLE ... ADD COLUMN` is supported, and even then, only for
a subset of its behaviors.

Release note: None